+++ Library/src/main/java/mezz/jei/library/runtime/JeiHelpers.java
@@ -20,7 +20,7 @@ import org.jetbrains.annotations.Nullable;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Optional;
-import java.util.stream.Stream;
+import java.util.concurrent.Executor;
 
 public class JeiHelpers implements IJeiHelpers {
 	private final GuiHelper guiHelper;
@@ -32,6 +32,7 @@ public class JeiHelpers implements IJeiHelpers {
 	private final IVanillaRecipeFactory vanillaRecipeFactory;
 	private final IIngredientVisibility ingredientVisibility;
 	private final IPlatformFluidHelper<?> platformFluidHelper;
+	private final Executor clientExecutor;
 	private @Nullable Collection<IRecipeCategory<?>> recipeCategories;
 
 	public JeiHelpers(
@@ -41,8 +42,7 @@ public class JeiHelpers implements IJeiHelpers {
 		IFocusFactory focusFactory,
 		IColorHelper colorHelper,
 		IIngredientManager ingredientManager,
-		IVanillaRecipeFactory vanillaRecipeFactory,
-		IIngredientVisibility ingredientVisibility
+		Executor clientExecutor
 	) {
 		this.guiHelper = guiHelper;
 		this.stackHelper = stackHelper;
@@ -53,6 +53,7 @@ public class JeiHelpers implements IJeiHelpers {
 		this.vanillaRecipeFactory = vanillaRecipeFactory;
 		this.ingredientVisibility = ingredientVisibility;
 		this.platformFluidHelper = Services.PLATFORM.getFluidHelper();
+		this.clientExecutor = clientExecutor;
 	}
 
 	public void setRecipeCategories(Collection<IRecipeCategory<?>> recipeCategories) {
@@ -130,12 +131,7 @@ public class JeiHelpers implements IJeiHelpers {
 	}
 
 	@Override
-	public IVanillaRecipeFactory getVanillaRecipeFactory() {
-		return vanillaRecipeFactory;
-	}
-
-	@Override
-	public IIngredientVisibility getIngredientVisibility() {
-		return ingredientVisibility;
+	public Executor getClientExecutor() {
+		return clientExecutor;
 	}
 }
